N21-2500 moved fetch service from school license module
