+++ apps/server/src/infra/vidis-client/vidis-client.adapter.ts
@@ -47,6 +47,39 @@ export class VidisClientAdapter {
 		}
 	}
 
+	public async getOfferItemsBySchoolName(mediaSource: MediaSource, schoolName: string): Promise<OfferDTO[]> {
+		if (!mediaSource.vidisConfig) {
+			throw new MediaSourceVidisConfigNotFoundLoggableException(mediaSource.id, MediaSourceDataFormat.VIDIS);
+		}
+
+		const { vidisConfig } = mediaSource;
+		const api = IDMBetreiberApiFactory(
+			new Configuration({
+				basePath: vidisConfig.baseUrl,
+			})
+		);
+
+		try {
+			const axiosResponse: AxiosResponse<PageOfferDTO> = await api.getActivatedOffersBySchool(
+				schoolName,
+				undefined,
+				undefined,
+				{
+					headers: { Authorization: this.buildAuthHeaderValue(vidisConfig) },
+				}
+			);
+			const offerItems: OfferDTO[] = axiosResponse.data.items ?? [];
+
+			return offerItems;
+		} catch (error: unknown) {
+			if (isAxiosError(error)) {
+				throw new AxiosErrorLoggable(error, 'VIDIS_GET_OFFER_ITEMS_FOR_SCHOOL_FAILED');
+			} else {
+				throw error;
+			}
+		}
+	}
+
 	private buildAuthHeaderValue(vidisConfig: MediaSourceVidisConfig): string {
 		const decryptedUsername = this.encryptionService.decrypt(vidisConfig.username);
 		const decryptedPassword = this.encryptionService.decrypt(vidisConfig.password);
