+++ apps/server/src/modules/school-license/school-license.module.ts
@@ -1,18 +1,17 @@
 import { LoggerModule } from '@core/logger';
-import { EncryptionModule } from '@infra/encryption';
+import { VidisClientModule } from '@infra/vidis-client';
 import { MediaSourceModule } from '@modules/media-source/media-source.module';
 import { SchoolModule } from '@modules/school';
 import { Module } from '@nestjs/common';
 import { MEDIA_SCHOOL_LICENSE_REPO } from './repo';
 import { MediaSchoolLicenseMikroOrmRepo } from './repo/mikro-orm/media-school-license.repo';
-import { MediaSchoolLicenseFetchService, MediaSchoolLicenseService } from './service';
+import { MediaSchoolLicenseService } from './service';
 
 @Module({
-	imports: [MediaSourceModule, SchoolModule, LoggerModule, EncryptionModule],
+	imports: [MediaSourceModule, SchoolModule, LoggerModule, VidisClientModule],
 	providers: [
 		{ provide: MEDIA_SCHOOL_LICENSE_REPO, useClass: MediaSchoolLicenseMikroOrmRepo },
 		MediaSchoolLicenseService,
-		MediaSchoolLicenseFetchService,
 	],
 	exports: [MediaSchoolLicenseService],
 })
