+++ apps/server/src/modules/school-license/service/index.ts
@@ -1,2 +1,1 @@
 export { MediaSchoolLicenseService } from './media-school-license.service';
-export { MediaSchoolLicenseFetchService } from './media-school-license-fetch.service';
